server/requirements.txt
flask
flask-sqlalchemy
androguard
pillow